LayoutTests/platform/chromium-mac/fast/dom/nodesFromRect-basic-expected.txt

CONSOLE MESSAGE: line 14: Uncaught TypeError: Object [object Object] has no method 'nodesFromRect'

PASS successfullyParsed is true

TEST COMPLETE
